requirements.txt
Flask==1.1.1
requests==2.20.0
plaid-python>=5.0.0, < 6.0.0
discord
